LPS-89891 Adds new management toolbar class for layout page templates
